Put Absorber demo into it's own branch.

+++ src/ModelPackage/IGizmo.java
@@ -15,7 +15,7 @@ public interface IGizmo {
 
     int getEndyPosition();
 
-    int getRotation();
+    double getRotation();
 
     Color getColour();
